Add a testing harness This commit finishes up the reworking of the Makefile, I think, and introduces a testing harness that allows new tests to be created and added easily.  New tests are added by adding a .dfn file containing the program to test and a _r.c file of the same suffix that is used to verify the results. The testing harness links these two files together with the driver and the runtime to produce an executable that should return an EXIT_SUCCESS if the run_test() program, called with no arguments, returns zero. Otherwise it will exit with a failure and report it on standard error. There are three basic tests right now, two of which work and pass. The third test fails because of some API changes that need to be made as well as a runtime that does not yet exist.
The makefile reworking is done in the style of the paper 'Recursive
Make Considered Harmful' and uses a single global Makefile that
includes child makefiles to get all of its targets. We should
be gerating a precise, complete dependency graph, so any missing
dependencies are bugs, and should be fixed. Moreover, we try to ensure
that we only recompile or reprocess what we need on demand.