Remove Va from Function Enumeration Set
It turns out that we don't need this to understand how Functions work,
because in every case they are subsumed by the E stimuli. This change
also allows us to refactor Va to be a variable expression, rather than
just a single variable, exposing the possibility for nested assignment.
This necessitates the introduction of a Vnu stimuli which will be used
with E.