Distinguish between single and nested bound array
If we are going to handle selective assignment correctly, then we need
to have the precision to distinguish between nested and single array
bound variables. Selective assignment admits only a single array-bound
variable, while in general we want to have any nested array-bound
variables for our use. This commit also includes some adjustments that
simplify and further reduce the follow-up on the N enumeration for
expressions.