Update Va [ +
It turns out that using a variable in an index is quite different than
using a number in the index. This is obvious in a few cases, such as
when we want to assign to it. However, it also reveals some interesting
things about strand assignment. Namely, we can have something like Va [
] Vna ← and that makes sense. This will require a different handling of
the Va and Vna enumerations, which is coming forthwith.