Update top-level to have new stimuli
To handle the explicit naming of the function expressions, we now have
to deal with that in the top-level as well as in some of the other
contexts. This initial update adds the stimuli and gives reasonable
values to the tables that were already enumerated, leaving two
additional tables for follow-up.