Finish the reworking of Top-level Enumerations
This completes the reworking of the top-level enumeration to handle the
difference in the way we handle named functions. Hopefully this is
correct, as I went through this pretty quickly, and while I am
relatively sure that it is correct, it could do with another go around
at some point.