Begin Transition of Operator Primitive Handling
Due to the interesting nature of the way that primitive operators work,
I've decided that it is just easier to be explicit about each individual
one, than to try to understand them as a class. This means that I have
to move away from using Om and Od and into use the individual
primitives. This is the first of a series of commits that aims to do
just that. This commit updates the tables of stimuli to eliminate the Om
and Od stimuli and replace them with more specific versions, as well as
updates the stimuli set of the Function Expressions.