Fix errors in previous enumerations
This fixes problems with the handling of the / and \ operators (which
can also be functions) as well as handling of Ea ( in the Func. Expr.
Enumerations. We are going to have to be careful with Ea (. We would
like to delete it and consider it a syntax error, but in Dyalog, some
operators are allowed to have parentheses around them, and so this makes
Ea ( a valid history to consider.