Update handling of operators and Function vars.
This commit makes some changes to the way that we handle operators and
functions and especially their variables. The new Vfo stimuli captures
what was previously captures in Vf, Vom, and Vod. Vom and Vod are now
gone in favor of the Vo[op] stimuli. The changes to the Functions and
Top-level enumerations reflect these alterations, while the Function
Expressions enumerations do not yet reflect these changes.

These changes have the net additional benefit of allowing for operators
to be named and bound, which builds on the previous patches to use the
Feo stimuli. All the enumerations besides Function Expression
enumerations now reflect this new capacity. Much more work must go into
making this happen for the Function Expressions enumerations, so this
will arrive as a separate commit.