Update Function Expressions for Operator Values
This commit finally brings operators into the fore with proper operator
responses and the proper use of the Vo stimuli for binding of operators
to names. This updates only the previously completed enumerations in the
Function Expressions set of enumerations. Progress can now continue on
the rest of the enumerations.