Major adjustments to enumeration of Expressions
This is the first commit moving to a canonical analysis of the
Expression Sequences, but during this analysis a number of duplicate
states were discovered, which caused an audit of the expression tables,
which revealed some problems with handling of the selective assignments
and of the atomic responses under indexing. These problems have been
addressed in this commit, and a few new tables added to accommodate the
changes.