Add Black Box Definition for ( and more
This commit makes some significant changes in the Function Expression
enumeration to allow for better handling of parentheses in the Fn case.
It also identifies locations of redundancy in the state space and
removes them. On the whole, this is a slightly more complex graph than
the previous, but this is because of the addition of more handling of
parentheses for Fn, despite the reduction in states for handling
parentheses with Da and Fn \.