Begin Increment Test Plan
This commit includes changes in the Function Specification that makes it
easier to write the Increment Test Plan by enumerating the Canonical
States in the Function Specification rather than requiring this
enumeration in the Increment Test Plan. The Increment Test Plan
currently includes the Usage Model Overview, but nothing more.