Update Usage Model representation
It doesn't quite make sense to have all of this duplicate information
when writing each usage model, as the graphs of the state space and the
edges doesn't actually change, but only their probabilities. We also
need to make sure that we account for multiple increments as each of
these models will be different depending on what increment we are
working on. All of this combines to make a compelling case for splitting
up the (Prob, Target) pair into two different descriptions, one for the
more static targets, and one for the much more changeable Probabilities.
This commit incorporates these ideas, and starts work on the actual
elements for making it easier to work with these models.