Cleanup ConvertFree handling of scopes
In the previous code there was a type (⍺ → ⍺⍺) that should not
have been there. Additionally, when entering a new function scope,
the stack frame of the existing environments must increase one to
make room for the local scopes. Finally, it doesn't make sense to
keep around two copies of the scope, when the only time we care about
the "total function scope" passed around in vis as a left operand is
when we enter a new scope and push all the previous scope onto the
existing environment, so that we can treat those values as already
being in scope. In this case, we only want to add the values that
are in the scope right above us, as we presume that all the other
entries are already in the environment. Thus, in the left operand to
vis we can safely keep only the current local scope there and don't
need to continue beyond that, while the main env argument in the
right argument to vis will maintain all of the rest of the information.