Major adjustments, bugfixes, and passes
After a verification of code, there were a number of issues that
were corrected. Additionally, two passes now bite the dust as
unnecessary and unproductive. ConvertFree becomes AnchorVars now,
taking the place of the AssignVars and Allocate passes. We remove
the DropUnmd pass as not helping very much.

But fixes throughout the passes and some reordering to make things
work better. AnchorVars should now be complete.

This basically completes step 10 of Increment 5.