Update GenFunc
Make sure that GenFunc adds the appropriate allocation to each
function and pass that new environment along as an extra operand
element for the GenExpr and GenCond functions. Also adjust the
type of a function to match the official docs now with the
Software Architecture in line with the code. This means that
GenFuncType now takes an actual depth which is used to calculate
the arguments that it takes.