Implement GenFnAlias and Factor out Function Creation
The GenFnAlias function now has an implementation, but it
required some rework to ensure that it could create the aliases.
This is because we don't know the depth of a function at the
time that we reference it, which means that it helps to have a
table of this already available. If there were a way in LLVM
to get the Type of a function automatically, I wouldn't need to
do this, but I can't see an obvious way of handlig this, so I
just create a new type from the depth each time we need to do
the aliasing.

Other than that, we need to create a function body template
in both the GenFunc and the GenInit case, so I've abstracted
this out into its own function and made sure not to create an
allocation frame if the allocation size is zero. This avoids an
undefined behavior in the LLVM system and also allows for cleaner
code generation in the GenInit case.