New GenInit Function
This function handles the initialization of globals, since
LLVM doesn't allow arbitrary code to be run at initialization
time, so we encapsulate this in a function which does this work
for us. This is a special function, in that it mutates the global
variables, which means that we cannot just put it into the
form of a regular function and let the regular function generators
handle it.

To accomplish all of this, we also needed to put in a reference
to a function which will create the function declarations for all
of the runtime code, and we incorporated a few other changes here
and there to clean things up as we found useful abstractions to
insert and use.