Update GenFunc for correct return handling
The change in the function signatures also requires a change
in the way that function results feed back out of the function.
In this case, we need to make sure that the res parameter contains
the correct result, that we return a zero error code, and that
we cleanup the environment before returning.

All of this required that we pass along the size of the environment
with the local environment to ease calling the cleanup function. It
also requires a different kind of MkRet function, and we are
deprecating the MkEmptyReturn function. Finally, there is a new
function needed for getting the res argument from the function.