Adapt GenCond to new conditions
This commit adapts GenCond so that it can handle our new
forms. This required a bit of work, but ends up with a cleaner
GenCond than we had before.

Firstly, we fixed a bug in the vex function of LiftBound.

Secondly, we adjust LiftBound to ensure that only atomic variable
references occur in the Condition test expressions, so that a
Condition node does not change its environment.

Next, we lift out the work of generating a block of code in a
function, which is the same in a function body as well as the
Condition consequent, now that we allow for multiple expresions in
the consequent of a condition node. This allows us to abstract out
that work into the same place, which ensures that we keep a consistent
semantics and all of our new function behaviors should be preserved
both in consequent branches as well as in function bodies.

We add a help function LookupExpr that helps to get the right
bindings from a value in the name bindings. First it tries what is
in the environment, and if that is not found, it looks in the global
module bindings list.

We keep most of the logic in GenCond the same, as nothing much has
changed here, but we abstract out the work of building the consequent
block to our GenFnBlock helper now. Additionally, we do the work of
lookup using LookupExpr. We still assume that we have the right form
and we don't check any of that.