Improve the LookupExpr function
Looking up an expression is actually a little more complicated.
There are three cases we can have when we need to lookup the variable,
in the first the variable already exists in the environment. In
this case we just return it. If it is not already in the environment,
it could be in the global space, but only if the environment reference
of the Variable attribute extends beyond the number of environments
passed to the function. In the most complicated case, we need to
create a reference to the variable in the appropriate frame and giving
it the appropriate name and inserting it into the bindings environment.

This commit fixes this up and makes the usage work out.