Cleanup GenExpr and friends
This makes a number of rather large and sweeping changes in order to
get a reasonable GenExpr. Firstly, it's been linearized more, so it reads
nicer. We also changed the way that we generate functions, so that
GenFunc doesn't generate aliases, and instead, a separate GenFnDec pass
before the GenFunc pass handles the declaration of the functions. This
allows us to do a lookup for a function name easily without requiring additional
information. We also make sure that we update the environment with the
assignments when generating expressions, but not before we lookup the
variables on the right, to ensure the appropriate ordering of things.