Fix addf in handling empty case
When adding functions, it's possible that addf is called each on an
empty vector, which leads to the fill element getting passed to the
addf function. Originally this was assumed to be a zero length vector,
but this is wrong. It is actually a n-length vector of spaces. We change
this code to handle this case correctly.