Cleanup the ModToNS function based on verification
This includes some adjustments to make sure that functions, when called,
appropriately handle their arguments, and also ensures that the runtime code
has available some of the functions necessary to do this. We also make sure
that some of the functions, clean_env and array_cp go into the runtime.