Fix up ConvertArray
We fix a standing issue as to the behavior of FFIGetDataInt and FFIGetShape.
They only have one output element, so they return the contents of that single
output element, not the a single vector containing that output. This required
some examination of the other code, which during execution to determine the
behavior of the ConvertArray function, revealed issues in the way that the FFI
for Dyalog works. Namely, we cannot assume that we can continue to use the values
that are passed into us. Instead, we need to make sure that we allocate these
afresh and copy data into them. This also fixes an issue with array_free, where
trying to free an array created by FFIMakeArray would crash because the pointers
into the shape and data elements would fails, since before they were just copied
from the Dyalog pointers. Now they are deep copied into freshly allocated
space, and using array_free() on such arrays works correctly.