Verify and adapt Parse
We go through the Parse function and rewrite/verify it to handle the
new style and to make it easier to read through the code. We have
adjusted a few things to favor the use of masks instead of index sets.
Much of the documentation has been eliminated in favor of terser versions,
while some of the important external documentation has been moved to
the Software Architecture.