Verify and fix ParseFunc
Simplify the ParseFunc procedure some, and fix an issue with
environment handling. Previously this function was not handling the
environment that was passed into it, and it was instead just taking
an empty environment as its starting point. instead, it now extends
the current environment with new bindings to ⍺ and ⍵ as it should.