Major cleanup and verification on ParseExpr
I am not sure what I was thinking when I wrote that thing the last
time, but it had a number of errors in it that are now fixed, and
the comments make it much easier to read. The code is slightly
more verbose, but much more direct and easier to understand.