Update the runtime linkages
The runtime written mostly by Tim Zakian isn't quite the way that
we initially thought it would be so we have adapted it to work.
Additionally, I have abstracted out the appropriate elements into
the top-level so that they can be shared and easily edited in
the future to add more primitives.