Rewrite the codfns_add function in preparation for more abstraction
This commit makes a number of changes that I think are better to
the runtime as a whole. It abstracts out the common scalar operations
into a series of smaller functions and one higher-order function
that is parameterized based on the scalar function operation performed
over a single scalar element. This makes is possible to see each
piece on a single screen and also reduces code duplication.

It relies on some good inlining though, and I'm not sure whether this
will be faster.