Initial sketch of additional functions to suppor blackscholes
None of these have support for floating point numbers right now
and they are intentionally very naive because of time constraints
and not knowing how the rest of these functions will evolve over
time. They make a lot of assumptions and are considered more of
prototypes or code for the benchmarks and early work than something
worthy of certification at the moment. They compile, and that's
all that matters at the moment.