Update handling of LiftBound
This update adjusts the LiftBound function by quite a lot. It
greatly shortens the code and makes it easier to follow, but it
also improves the invariants of the output AST. In particular, it
completely flattens the expressions now, and makes sure that all
unnamed expressions are variable reference, rather than calls
or references.

It further addresses an issue in the previous code which would
prevent the lifting functions from entering into function bodies
nested inside of expressions, which are not yet lifted at this
point.