Make statement oriented error handling a lot better
This uses a new combinator _aew (any ends with) that captures the
concept of a series of elements that should be terminated explicitly by
something else. This allows for better understanding of the errors in
cases where we have successfully parsed a number of statements, but
another fails to parse, but we have not reached the end of a statement
block as indicated by a terminator of some kind, such as a brace. The
result is that we get much better tracking of the actual parse error
that created the fault and can see this in the pass back through the
state to the final result.