Rewrite rd
the rd pass now is written as a train as well, using the new AST format
and adds the depth annotation to each of the nodes, not just to the
function nodes. We also change the rd to annotate with the depth
including self, not the depth not including self, to make the numbers
look prettier.