Add free variable types to S nodes, fixes #135
This one required some heavy lifting. Mostly in the form of a new compiler
pass that normalizes the shape of the `v` attribute. The `e` attribute then
takes its overall shape from that, and with some adjustments to the intervening
passes, it is possible to ensure that everything works according to spec.