Initial update to #224
This fixes up this issue by introducing support for Visual Studio and
making it the default compiler out of the box. Right now this creates
two different behaviors in the compilers on Linux and those on Windows
because the one on Windows just needs the name of the file and uses the
BUILD∆PATH option to tell where to put the files. This should be the new
form used in the Linux compilers as well, but it is not there yet. I
have fixed up a few tests to use this new system just enough to tell
that things are likely to work out, and to get something ready for
people to start playing with, but I will  not close the ticket until I
have the rest of the tests ported over to work.