Add commute operator
Part of #180; this fixes up the question of how to handle operators and
their operands in the general case, but does not allow for optimizations
of the scalar functions in cases where this could be of benefit.
However, it does provide a solid foundation for the rest of the
operators to be implemented on. Basically, it allows for the sdb scalar
database to be redone and repurposed as an extension to the fdb mixed
function database, permitting a scalar function to be used as a mixed
function without going through the standard scalar fusion generation.