Refactor Each Handling, Initial Reduction Support, More Cleanup
Part of #180, this moves work forward towards providing better and more
efficient handling of operators and functions, especially in the case of
flat arrays and scalar functions, but it also sets the stage for
reduction with an initial implementation of reduction that works over
scalar functions. It provides the mechanism to call mixed functions on
scalar arguments inside of a mxfn call, but this has some limitations,
as there is a weakness in the handling of mxfn and in the cal{m,d}
functions which could invalidate array data pointers during collection
which could cause seg faults. I don't have a solution to this right now,
but I am sure that one is technically possible. This problem is a part
of #236 and should be dealt with in that ticket.