Complete Reduction
Part of #180, this commit fixes up the type checker to handle operators
that have to deal with different arities compared to the way that they
call their functions, allowing for type indirections in the checker.
this fixes the error with handling mixed functions in reductions and
completes the implementation details for reduction.