Tested on Win 8.1.
Need to make the file more readable, with tabs. It is difficult to read now. 

Has been tested on Windows 8.1. I've used the 'q' and 'w' keys to store the left tack(⊣) and the right tack(⊢) respectively. This is in *addition* to the OEM_5 key having them(specific to UK keyboards only). Some keyboards do not have the OEM_5  keyboard, hence this liberty.