Start a crazy refactoring to unify all runtime code generation
This is a rather ambitious but long in coming start to a change to the
infrastructure for the code generators. Right now the code generators
fall into three basic camps: scalars, mixed, and operators. These
have three different interfaces and the functions that generate the
end code, while similar, are different i their APIs. In reality, this
is unnecessary and leads to excessive complication, such as the large
dispatch tables for scalars and mixed functions. There leads to lots
of indirection which makes it harder to know what is going on, and
in general encourages too much abstraction for not enough gain.

Instead, the new system relies on dispatching from a single symbol
table for everything, and calls an appropriate constructed function
that is given the entire node to generate from, to simplify the API.
Furthermore, we call specific generators for each unique combination
of input type information, so that we can specialie the generators
based on available type information without requiring complex control
flow in the generators themselves. This greatly simplifies the overall
architecture required, hopefully making the generators read more nicely
as dfns without the excess abstraction of functions like mxfn.

This is only a start, but we've already pushed E1 into this new structure
and demonstrated how it works with the `⊢` monadic case. The goal is to
greatly simplify the runtime code and eventually eliminate most of the
unnecessarily cmoplex and hard to understand logic in the runtime.