Refactoring, cleanup, new interface
This is a major change that disables the type-dispatch pass and relies
on the runtime instead. It simplifies the runtime by eliminating the
function entry code and replacing it with a few preprocessor macros. It
simplifies the interface and fixes the naming conflicts issues because
all exports now have suffixes and all internal names are either
generated or also have suffixes. Removes a lot of unnecessary code and
simplifies the runtime significantly.