New Year, Simplify Things
This restructures the compiler to allow for easier management for those
who might not have more sophisticated editors. This reduces the overall
burden. We are also moving to allow for a DWS workspace based deployment
as this seems like it will make acquiring and using the compiler easier,
since you are not required to use SALT.