Improve the parser by a wide margin
This expands the parser's ability to handle arbitrary operators much
better, as well as making sure that Bops (bracket operators) are handled
appropriately. The parser's performance has been greatly improved as
well because of the avoidance of reparsing on a number of elements. This
should complete the parsing element of handling internal bindings.