Remove BUILD∆PATH/BSO, simplify AF∆LIB
This is a simplification of the way that the compiler works. Now it will always store and declare compiled objects and meta files in your current working directory. This has the advantages that it is no longer necessary to ensure that you have a build directory. It also makes it easier to package and distribute Co-dfns compiled objects since you can simply put everything in a single folder and send it on its way. It will also be possible to manually move the shared objects/DLLs to another location in your search path, since there are no explicit paths used for the ⎕NA calls.

This should make it much easier to use the User-commands infrastructure to work with Co-dfns.