Adapt parser to new format
Parser now works off of interned symbols in the structure, uses an inverted table representation, and we've adjusted the binding node to fall under their own binding node type instead of kinds of E and F nodes.

The third value returned by the parser is the symbol table for the n field, while the N∆ constant serves as the symbol table for the t field.