Better assignment handling
Adjust the parser and generator to be more consistent in its application of operator and function behaviors for modified and bracket assignments, both in combination and separate from one another. This also allows some scalability for implementing selective assignment when the time comes, without any real cost to us, and makes sure that we don't introduce any crazy number of generator cases, while keeping the number of runtime functions down, as well.