Do what we think is right
Adapt to what we think is right with regards to the Dyalog documentation and wait for confirmation from Dyalog devs to change things if this isn't correct. Interesting to note that this case doesn't appear to be handled appropriately in APLX, APL2, or GNU APL either. They all appear to give LENGTH ERROR in all cases.