Initial axis operator support for reduction
This is the initial work on axis operator based reduction, which should simplify most of the reduction code, but also revealed an interesting bug that we are now testing for. Solution has been added for axis operator reduction, but needs to be introduced into the other code paths (which will be a part of the simplification of this code after the final tiling error is fixed).