Get test suite 10 working
This is really adjusting for the fact that we can't determine span-ness on the basis of rank anymore, and so we need to use the f field, but we were not updating the f field on most primitives. We should do this now, and it should handle our spanness question.