Implement Fa as a PEG grammar
This makes some changes to the way that F types are generated, in that we now use the full 3 3 2 2 encoding for function types instead of simplifying that down. There are some issues with parsing user-defined functions still, but the code is there to give a hint to that at least. There needs to be more work on ambiguous parsing and the proper way to encode that before we can get into that fully.