Improve error display, track source information in AST, internal EN and DM
This improve the display of errors and tracks source information in the AST, a precursor to handling the internal Fn/Fa source information tracking. Additionally, we create our own EN and DM values inside of the compiler to handle diagnostic messaging instead of trying to use the built-in ⎕DM and ⎕EN. Information is still populated into the ⎕DMX as we can.

#453