Make TK provide one token at a time
Instead of chaining TK and PS together, we subordinate TK so that it is spawned each time PS needs a new token. This is to enable us to further abstract the handling of expressions into a separate parsing module and allow TK to manage the abstraction layer a bit better.