Fix removing Z nodes, pass TEST 4
When removing Z nodes, we can undermine the sibling ordering invariants if we simply remove the Z node positions and link everything through the appended nodes, such as the E and A nodes that are added late.

Additionally, Z nodes may be highly nested as a result of excessive parentheses, so we need to make sure that we are linking deeply through a potential chain of Z nodes.

At this point in the parsing we can assume that Z nodes should only have a single child, which could be a Z node.

To properly eliminate Z nodes while retaining sibling ordering invariants we should instead move the child nodes into the Z node position that is highest in the chain, relinking grandchildren to the original Z node. At that point, the original child nodes and any spurious intermediate depth Z nodes can be removed normally.