Convert primitives to variables
Before we do the rest of the generation, we want to convert the primitive values into variables that are global. This requires that we also update the symbol table so that the names of the primitives are no longer APL symbols but the names that map to global runtime references in our runtime.

We need to do this in the code generator because we might have different name mappings and ways of handling this depending on the code generator, so we will leave the primitive node type until generation. However, we don't want to have to write special code generation code on primitives; rather, we want to be able to treat all of these references as some type of variable node, so we need to remove them before moving on through the rest of the generator.