We're not using slots and frames or counting strand or indexing anymore
Because we are going to need to use the standard C variable names with different scoping rules, we will no longer need to calculate frame depths or the like.  We also don't need to keep strand or indexing counts because we will keep variable style references as the children for the code generator.