Replace nodes with V0 when we lift and flatten expressions
Rather than pushing everything to the stack, we are now not pushing variables to the stack and leaving primitives as well. This is because we now consider these "reference" nodes rather than "operational" nodes.

When we lift operational nodes to the stack, we want to leave behind an indication that the result of that expression is on the stack, so we leave a V0 node in its place.