Declare functions at the top and don't prefix names
We don't need to prefix the names since we shouldn't have to have names of different types but the same root sitting around. We also want to make sure that we declare the function at the top prefix to ensure that other functions can create closures from the global function declaration.