Allow for G children which are out of order in their creation
The previous code operates under the assumption that the two children of any given G node are also the nearest to one another among the entire set of G node children. With changes to the parser, this is now not always the case. Instead, we sort the children by their parent to ensure that they are contiguous and then select the second child.