Split error functions into their own definitions at top-level
This is tricky because now we have some free references that we need to track and ensure work across all of our functions. This includes IN, pos, and end. We also have to make sure that `SIGNAL` and `SELECT` are both trad-fns so that they can capture the appropriate variables. Bah, humbug.