Parse arrays in a slightly more sophisticated manner
This method of parsing arrays handles strings better and also ensures a more consistent handling of simple arrays by delaying the evaluation of simple arrays until after we have fully removed the Z nodes. This means that mixed simple arrays that have parentheses in them will still be merged into the appropriate A1 nodes.

We also change the A1 node format to use a literal n field instead of nested N nodes. This unifies the handling of strings and numeric literal arrays into a single A1 node instead of having separate types to handle both of these. We also move to using the A7 node to handle stranded arrays.

Additionally, this introduces the idea of using A0 nodes to handle niladic calls to functions that will return an array value, such as ⎕ or ⍞, but also niladic variables. We currently do not check for niladic variables, but it is easy to add this at some point in the future.