Revert our Z node changes
The original Z node removal works the way it does because we cannot be sure that the children of a Z node are in the same order as the parents. Thus, instead of trying to reorder the children as appropriate, we delete the children and excess Z nodes, but keep the topmost Z node of any chain and convert it to the original salient child node, redirecting all of the grandchildren to it. This preserves Z ordering without needing to rearrange nodes.