Be more meticulous about how we retain cells
We can be much better about retention and release of cells. In particular, we only need to retain a cell when we bind it to a name or when we push it onto the stack. Referring to the name should not in general result in us retaining it further. Moreover, given this, we don't need to release a cell if it is bound to a name just because we reference it. Instead, we only need to release cells that we pop off the stack, since those are "gone". Likewise, we will need to retain cells when we create a closure out of them, but otherwise, we don't need to retain as much as we were doing.