Compute variable lookup on the basis of closure-based variable propagation
This mentally simplifies the overall system and also uses closures and closure creation to thread variables through the system. This makes it more simple to calculate free and local variables as well as ensuring that we know the locality of each variable at each stage up through the functions, which wasn't the case with the previous way of handling variable anchoring in the parser.