Refine and simplify the structure of scalar functions in the runtime
This new approach reduces the overall number of ibeams that we need to implement and has the particularly valuable effect of reducing the overheads involved in each underlying loop iteration for scalar functions operating over nested arrays. Generally, this design is at least as easy or more easy to implement and also more consistent with the design of the language, while promising to deliver better overall code generation.