Make allocation of data separate from allocation of array header
We are simplifying the core mk_array() function to only do the basic allocation. After that, you need to make sure to set the shape accordingly and also separately fill/allocate the data values.

This is done to allow for arrays to share data buffers. Some operations should be able to share data buffers rather than requiring that they allocate new data all the time. This will allow some operations to be very fast, and thus permit their use in the runtime code without incurring excessive overheads.

This also necessitates the introduction of a vrefc pointer that tracks the values reference count for host allocated data.