Make sure to zero the array when we allocate it on the host
We need to make sure that we zero our host data buffer. If we are in the process of populating that data buffer and then enter a failure state, when we release the array, we may need to release the partially filled buffer, so we want to make sure that we are releasing on either NULL values or real arrays, not junk data.