Make loc a pointer; point to global namespace before populating
This allows us to make better code in the runtime. The previous design assumed that `loc` would be filled up first, and then we would copy those values to the cdf_prim structure, which is the top-level namespace of the runtime. However, this doesn't actually work in practice if we want to access the primitives in their primitive names rather than using the runtime local names, since `cdf_prim` will not be populated until after we have fully populated the loc structure. Instead, we map the loc structure in all cases to be a pointer to the structure in question, which allows us to directly populate the global structure as we go, which allows us to access the right functions provided that we define them first in the runtime code.