Manually deal with functions that don't return a value
The DWA mechanism is not able to handle functions that don't return a value, so instead of trying to do this through the DWA system, we manage this ourselves. We do this by returning a -1 error code signaling that the function has no return value. We also need to ensure that we put some dummy value into the return slot so that DWA doesn't complain. We need to exit out early and adjust how we track the result Z to make sure that we don't set Z if we don't have a value to use.