Make error returns work in the presence of single references
The previous version of this code assumes that any statement that is executed within a dfn will set `err` appropriately, but this is not true. In the case of reference or passthrough values, we do not set the `err` status. This could cause issues. To fix this, we put the `err = -1` statement at the end of the statements, since it will be skipped if we have a viable return value, and initialize `err` explicitly.