Make internal assert a displayed syntax error
If we fail to parse things, which means ending up with Z nodes at the end of the parser with more than one child in it, then we should report those Z nodes as failing to parse. This is a pretty great quality of life improvement when things go wrong.