Move function lifting into PS
We need to lift functions before we replicate them for handling ambiguous parsing, so let's do that here. This has some fall out.

We decide to keep the p variable around and pass it out of PS.

We also now need to make sure that n's pointer values are appropriately updated.

To retain a connection between the original variable site and the newly lifted function, we make the function's n field point to the variable where it was originally located. This also requires that we adjust how we handle the namification of pointers.

This adjustment also requires that we be careful about how we handle the r field in TT now. We make a few tweaks to ensure that it flows all they way through the lexical stack as intended.

We take this opportunity to refactor marking mutable bindings a little bit and clean that up.